Conversational AI/Conversational AI/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-google-genai
faiss-cpu
python-dotenv
tavily-python
serpapi
requests
langchain-community